queues/blocking/src/test/java/ua/ds/concurrency/SingleLockLinkedBlockingQueueTest.java
package ua.ds.concurrency;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

class SingleLockLinkedBlockingQueueTest {

  private SingleLockLinkedBlockingQueue queue;

  @BeforeEach
  void setUp() throws Exception {
    queue = new SingleLockLinkedBlockingQueue();
  }

  @Test
  void addOneItem() throws Exception {
    queue.enqueue(10);

    assertThat(queue.deque(), is(10));
  }

  @Test
  void addManyItems() throws Exception {
    queue.enqueue(10);
    queue.enqueue(20);
    queue.enqueue(30);

    assertThat(queue.deque(), is(10));
    assertThat(queue.deque(), is(20));
    assertThat(queue.deque(), is(30));
  }

  @Test
  void addOneByOne() throws Exception {
    queue.enqueue(10);

    assertThat(queue.deque(), is(10));

    queue.enqueue(20);

    assertThat(queue.deque(), is(20));

    queue.enqueue(30);

    assertThat(queue.deque(), is(30));
  }
}
